Refactor: Controller, Service 로직 refactoring #2   - Transactional rollbackFor option    적용   - AuthorizationFilter excludepattern  적용
