+++ plot-server/build.gradle
@@ -60,6 +60,7 @@ dependencies {
 	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-mail', version: '2.6.3'
 
 
+
 }
 
 tasks.named('test') {
