+++ plot-server/src/main/java/com/plot/plotserver/controller/EmailController.java
@@ -5,12 +5,12 @@ package com.plot.plotserver.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.plot.plotserver.domain.EmailTmp;
 import com.plot.plotserver.domain.Message;
-import com.plot.plotserver.domain.RefreshToken;
 import com.plot.plotserver.dto.request.email.EmailCodeReqDto;
 import com.plot.plotserver.dto.request.email.EmailRequestDto;
 import com.plot.plotserver.exception.email.EmailCodeExpiredException;
 import com.plot.plotserver.exception.email.EmailCodeMismatchException;
 import com.plot.plotserver.exception.email.EmailCodeSendingFailureException;
+import com.plot.plotserver.service.EmailSendingService;
 import com.plot.plotserver.service.EmailService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -32,29 +32,19 @@ import java.util.Optional;
 @Slf4j
 public class EmailController {
 
+
     private final EmailService emailService;
 
 
     @PostMapping("login/mailConfirm")//인증 코드 전송
-    @Transactional
     public void mailConfirm(HttpServletResponse response,@RequestBody EmailRequestDto emailDto) throws MessagingException, IOException {
 
 
         ObjectMapper om = new ObjectMapper();
         response.setContentType(MediaType.APPLICATION_JSON.toString());
 
-        try {
-            String authCode = emailService.sendEmail(emailDto.getEmail());//인증번호 전송과, emailTmp 저장을 transaction으로 묶음.
 
-            EmailTmp emailTmp = EmailTmp.builder()
-                    .userEmail(emailDto.getEmail())
-                    .code(authCode)
-                    .createdAt(LocalDateTime.now())
-                    .build();
-            emailService.save(emailTmp);
-        } catch (Exception e) {
-            throw new EmailCodeSendingFailureException("인증번호 전송에 실패하였습니다.");
-        }
+        String authCode = emailService.sendMailCodeAndSave(emailDto);
 
         Message message = Message.builder()
                 .status(HttpStatus.OK)
@@ -69,27 +59,14 @@ public class EmailController {
         ObjectMapper om = new ObjectMapper();
         response.setContentType(MediaType.APPLICATION_JSON.toString());
 
-        Optional<EmailTmp> findUserEmail = emailService.findByUserEmail(emailCodeReqDto.getEmail());
+       emailService.mailCodeAuthenticate(emailCodeReqDto);
 
-        if (!findUserEmail.isPresent()) {
-            throw new EmailCodeExpiredException("인증번호가 만료되었습니다.! 다시 전송해주세요.");
-        } else if (findUserEmail.isPresent()) {
-                log.info("code={}",findUserEmail.get().getCode());
-            if (!emailCodeReqDto.getCode().equals(findUserEmail.get().getCode())) {//인증번호 틀림
-                throw new EmailCodeMismatchException("이메일 코드 틀림. 다시 입력해주세요");
-            }
-            else{
-                Message message = Message.builder()//정상 응답
-                        .status(HttpStatus.OK)
-                        .message("success")
-                        .build();
-                om.writeValue(response.getOutputStream(), message);
-            }
-
-        }
+       Message message = Message.builder()//정상 응답
+               .status(HttpStatus.OK)
+               .message("success")
+               .build();
+       om.writeValue(response.getOutputStream(), message);
 
     }
-
-
 
 }
