+++ plot-server/src/main/java/com/plot/plotserver/repository/EmailTmpRepository.java
@@ -17,4 +17,5 @@ public interface EmailTmpRepository extends JpaRepository<EmailTmp, Long> {
 
 
 
+
 }
