+++ plot-server/src/main/java/com/plot/plotserver/security/auth/JwtAuthorizationFilter.java
@@ -17,6 +17,8 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import org.springframework.security.web.util.matcher.RequestMatcher;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
@@ -25,6 +27,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -41,9 +44,20 @@ public class JwtAuthorizationFilter extends BasicAuthenticationFilter {
     // 1. 액세스토큰을 담고 있는 쿠키를 확인
     // 2. 액세스토큰이 유효하다면 -> 인증된 객체 저장하고 doFilter, 그렇지 않다면 -> 리프레스토큰 검사
     // 3. DB에서 리프레시토큰 조회. 리프레시 토큰이 유효하다면 -> 새로운 액세스토큰 발급, 그렇지 않다면 -> 인증된 객체를 저장하지 않고 doFilter
+
+    private final List<RequestMatcher> excludedUrlPatterns = Arrays.asList(//이 필터 적용 안할 url 지정
+            new AntPathRequestMatcher("/login/mailConfirm"),
+            new AntPathRequestMatcher("/login/authenticate")
+            // 추가적인 URL 패턴을 필요에 따라 설정
+    );
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException, com.plot.plotserver.exception.user.UserPrincipalNotFoundException {
 
+
+        if (isExcludedUrl(request)) {
+            filterChain.doFilter(request, response);//이 필터 스킵. 다음꺼 실행.
+            return;
+        }
         Cookie cookie = null;
         try {
             cookie = Arrays.stream(request.getCookies())
@@ -137,5 +151,10 @@ public class JwtAuthorizationFilter extends BasicAuthenticationFilter {
         // 인가 처리가 정상적으로 완료된다면 Authentication 객체 생성
         Authentication authentication = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
         SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+
+    private boolean isExcludedUrl(HttpServletRequest request) {
+        String requestUri = request.getRequestURI();
+        return excludedUrlPatterns.stream().anyMatch(pattern -> pattern.matches(request));
     }
 }
